Finished openSNP Parser
This implementation goes through all user IDs, contained in the file names on the server, and parses their phenotypic data, if the had entered some on the openSNP website. The result of the script, is a .csv file, containing all users who entered phenotypic data, their expression and a line containing everything that was entered, to make sorting easier.